requirements.txt
numpy
matplotlib.pyplot
random
argparse
pandas
os
sys